feat: modify the NearbyActivity
