+++ src/app/page.tsx
@@ -1,4 +1,3 @@
-import { getActivitiesByFilter } from '@action/activity';
 import CallToActionSection from '@components/CallToActionSection';
 import CommentSection from '@components/CommentSection';
 import HeroSection from '@components/HeroSection';
@@ -14,17 +13,7 @@ const Home = () => {
     <>
       <HeroSection className="mx-auto xl:mt-14" />
       <SearchBar className="mx-auto xl:mt-2" />
-      <NearbyActivity
-        getNearByActivities={getActivitiesByFilter(
-          {
-            lat: '121.5598',
-            lng: '25.09108',
-            limit: '6',
-          },
-          { cache: 'no-store' }
-        )}
-        className="mx-auto mt-24 xl:mt-36"
-      />
+      <NearbyActivity className="mx-auto mt-24 xl:mt-36" />
       <TopCategoryMenu className="mx-auto mt-24 xl:mt-36" />
       <RecommendActivity className="mx-auto mt-24 xl:mt-36" />
       <CallToActionSection className="mx-auto mt-24 xl:mt-36" />
