+++ src/components/NearbyActivity/index.ts
@@ -1,0 +1,3 @@
+import NearbyActivity from './NearbyActivity';
+
+export default NearbyActivity;
